refactor: ViewController 클래스 삭제 및 ViewController 설정 추가
