+++ src/main/java/roomescape/config/WebConfiguration.java
@@ -1,0 +1,20 @@
+package roomescape.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfiguration implements WebMvcConfigurer {
+
+    @Override
+    public void addViewControllers(ViewControllerRegistry registry) {
+        registry.addViewController("/").setViewName("index");
+        registry.addViewController("/login").setViewName("login");
+        registry.addViewController("/reservation").setViewName("reservation");
+        registry.addViewController("/admin").setViewName("admin/index");
+        registry.addViewController("/admin/reservation").setViewName("admin/reservation-new");
+        registry.addViewController("/admin/time").setViewName("admin/time");
+        registry.addViewController("/admin/theme").setViewName("admin/theme");
+    }
+}
