+++ src/main/java/roomescape/controller/ThemeController.java
@@ -1,4 +1,4 @@
-package roomescape.controller.api;
+package roomescape.controller;
 
 import java.net.URI;
 import java.util.List;
