--- src/main/java/roomescape/controller/view/AdminController.java
@@ -1,30 +1,0 @@
-package roomescape.controller.view;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/admin")
-public class AdminController {
-
-    @GetMapping
-    public String index() {
-        return "admin/index";
-    }
-
-    @GetMapping("/reservation")
-    public String reservation() {
-        return "admin/reservation-new";
-    }
-
-    @GetMapping("/time")
-    public String time() {
-        return "admin/time";
-    }
-
-    @GetMapping("/theme")
-    public String theme() {
-        return "admin/theme";
-    }
-}
