--- src/main/java/roomescape/controller/view/MemberController.java
@@ -1,20 +1,0 @@
-package roomescape.controller.view;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/")
-public class MemberController {
-
-    @GetMapping
-    public String index() {
-        return "index";
-    }
-
-    @GetMapping("reservation")
-    public String reservation() {
-        return "reservation";
-    }
-}
