+++ src/test/java/roomescape/controller/ReservationControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.api;
+package roomescape.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
